requirements.txt
google-generativeai
google-cloud-aiplatform
google-adk